requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
